christ. CodeBlocks is retarded.
also, lots of crap. I lost track of it.

+++ types.h
@@ -22,10 +22,10 @@
 typedef unsigned char       u8;
 typedef unsigned short      u16;
 typedef unsigned int        u32;
-typedef unsigned long int   u64;
+typedef unsigned long long int   u64;
 typedef signed char         s8;
 typedef signed short        s16;
 typedef signed int          s32;
-typedef signed long int     s64;
+typedef signed long long int     s64;
 
 #endif // TYPES_H
